Avoid recursive overrides
Recursive overrides happen if the extension was already loaded when this plugin loads. This leads to overriding the extension as part of the initial batch, which is fine. However, if the extension is loaded again using the overridden `load_extension`, the original function is lost and a recursive override takes its place.